Refactor Attachments app components (#1235)
