+++ ecm-wcm-extension/src/main/resources/locale/attachmentsSelector/attachments_en.properties
@@ -5,7 +5,7 @@ attachments.list=Attachment(s)
 attachments.list.empty=No attachment(s) yet
 attachments.view.all=View all attachments
 attachments.drawer.dropOrPaste=Drop files\u00A0or paste images here
-attachments.drawer.uploadOrdrop=Upload or Drag and drop your file(s) here
+attachments.drawer.uploadOrDrop=Upload or Drag and drop your file(s) here
 attachments.drawer.upload=Upload manually
 attachments.drawer.or=or
 attachments.drawer.maxFileSize=Maximum file size : {0} MB
