+++ ecm-wcm-extension/src/main/webapp/attachments-app/components/AttachmentsDrawer.vue
@@ -6,153 +6,54 @@
       right>
       <template slot="title">
         <div class="attachmentsDrawerHeader">
-          <v-btn
-            v-if="entityId && entityType || showDocumentSelector"
-            class="backButton"
-            icon
-            small
-            dark
-            outlined
-            @click="showDocumentSelector ? toggleServerFileSelector() : closeAttachmentsAppDrawer()">
-            <i class="uiIconBack"></i>
-          </v-btn>
           <span>{{ drawerTitle }}</span>
         </div>
       </template>
       <template slot="content">
-        <div v-show="!showDocumentSelector" class="attachmentsContent">
-          <div class="d-flex align-center">
-            <v-subheader class="text-sub-title pl-0 d-flex">{{ $t('attachments.upload') }}</v-subheader>
+        <div class="attachmentsContent pt-0 pa-5">
+          <attachments-upload-input
+            :attachments="attachments"
+            :max-files-count="maxFilesCount"
+            :max-files-size="maxFileSize"
+            :current-drive="currentDrive"
+            :path-destination-folder="pathDestinationFolder"
+          ></attachments-upload-input>
+          <attachments-uploaded-files
+            :attachments="attachments"
+            :schema-folder="schemaFolder"
+          ></attachments-uploaded-files>
+
+          <!-- Select From Drives Disabled for the moment -->
+          <!--<div class="d-flex align-center">
+            <v-subheader class="text-sub-title pl-0 d-flex">Platform Documents</v-subheader>
             <v-divider></v-divider>
           </div>
-          <div class="multiploadFilesSelector">
-            <div id="DropFileBox" ref="dropFileBox" class="dropFileBox theme--light" aria-controls @click="uploadFile">
-              <i class="uiIconEcmsUploadVersion uiIcon32x32"></i>
-              <v-subheader class="text-sub-title ml-3 d-none d-sm-flex" href="#" rel="tooltip" data-placement="bottom">
-                {{ $t('attachments.drawer.uploadOrdrop') }}
-              </v-subheader>
-            </div>
-            <!-- Select From Drives Disabled for the moment -->
-            <!--<div class="d-flex align-center">
-              <v-subheader class="text-sub-title pl-0 d-flex">Platform Documents</v-subheader>
-              <v-divider></v-divider>
-            </div>
-            <div class="lastContent d-flex align-center justify-center">
-              <a title="Select on server" class="uploadButton d-flex align-center" href="#" rel="tooltip" data-placement="bottom"
-                 @click="toggleServerFileSelector()">
-                <i class="uiIcon32x32FolderDefault uiIcon32x32LightGray"></i>
-                <v-icon color="#fff" x-small class="iconCloud">cloud</v-icon>
-                <span class="text colorText">{{ $t('attachments.drawer.existingUploads') }}</span>
-              </a>
-            </div>-->
-            <div class="fileHidden d-none">
-              <input ref="uploadInput" class="file" name="file" type="file" multiple="multiple" style="display:none"
-                     @change="handleFileUpload($refs.uploadInput.files)">
-            </div>
-            <transition name="fade">
-              <div v-show="fileSizeLimitError" class="sizeExceeded alert alert-error">
-                <i class="uiIconError"></i>
-                {{ $t('attachments.drawer.maxFileSize.error').replace('{0}', maxFileSize) }}
-              </div>
-            </transition>
-            <transition name="fade">
-              <div v-show="filesCountLimitError" class="countExceeded alert alert-error">
-                <i class="uiIconError"></i>
-                {{ $t('attachments.drawer.maxFileCount.error').replace('{0}', maxFilesCount) }}
-              </div>
-            </transition>
-            <transition name="fade">
-              <div v-show="sameFileError" class="sameFile alert alert-error">
-                <i class="uiIconError"></i>
-                {{ sameFileErrorMessage }}
-              </div>
-            </transition>
-          </div>
+          <div class="lastContent d-flex align-center justify-center">
+            <a title="Select on server" class="uploadButton d-flex align-center" href="#" rel="tooltip" data-placement="bottom"
+               @click="toggleServerFileSelector()">
+              <i class="uiIcon32x32FolderDefault uiIcon32x32LightGray"></i>
+              <v-icon color="#fff" x-small class="iconCloud">cloud</v-icon>
+              <span class="text colorText">{{ $t('attachments.drawer.existingUploads') }}</span>
+            </a>
+          </div>-->
 
-          <div class="uploadedFiles">
-            <div class="attachments-list d-flex align-center">
-              <v-subheader class="text-sub-title pl-0 d-flex">{{ $t('attachments.drawer.title') }} ({{ attachments.length }})
-              </v-subheader>
-              <v-divider></v-divider>
-            </div>
-            <div v-if="attachments.length === 0" class="no-files-attached d-flex flex-column align-center text-sub-title">
-              <div class="d-flex pl-6 not-files-icon">
-                <i class="uiIconAttach uiIcon64x64"></i>
-                <i class="uiIconCloseCircled uiIcon32x32"></i>
-              </div>
-              <span>{{ $t('no.attachments') }}</span>
-            </div>
-            <div v-if="attachments.length > 0" class="destinationFolder mb-4 ml-5">
-              <div v-if="!displayMessageDestinationFolder" :title="schemaFolder[0]" class="drive" rel="tooltip" data-placement="top">{{ schemaFolder[0] }}
-              </div>
-              <div v-if="showDestinationPath && !displayMessageDestinationFolder" class="folderLocation">
-                <div v-if="schemaFolder.length > 1" class="folder">
-                  <div><span class="uiIconArrowRight colorIcon"></span></div>
-                  <div :title="schemaFolder[1]" :class="schemaFolder.length === 2 ? 'active' : '' " class="folderName"
-                       rel="tooltip" data-placement="top">{{ schemaFolder[1] }}
-                  </div>
-                </div>
-                <div v-if="schemaFolder.length === 3" class="folder">
-                  <div><span class="uiIconArrowRight colorIcon"></span></div>
-                  <div :title="schemaFolder[2]"
-                       :class="schemaFolder[2] !== 'Activity Stream Documents' ? 'path' : 'active' " class="folderName"
-                       rel="tooltip" data-placement="top">{{ schemaFolder[2] }}
-                  </div>
-                </div>
-                <div v-if="schemaFolder.length > 3" class="folder">
-                  <div><span class="uiIconArrowRight colorIcon"></span></div>
-                  <div :title="schemaFolder[2]" class="folderName" rel="tooltip" data-placement="top">...</div>
-                </div>
-                <div v-for="folder in schemaFolder.slice(schemaFolder.length-1,schemaFolder.length)"
-                     v-show="schemaFolder.length > 3" :key="folder" class="folder">
-                  <div><span class="uiIconArrowRight colorIcon"></span></div>
-                  <div :title="folder"
-                       :class="schemaFolder[schemaFolder.length - 1] === folder && schemaFolder[schemaFolder.length - 1].length < 11 ?'active' : 'path'"
-                       class="folderName" rel="tooltip" data-placement="top">{{ folder }}
-                  </div>
-                </div>
-              </div>
-              <div v-if="displayMessageDestinationFolder" class="messageDestination">
-                <p :title="$t('attachments.drawer.destination.attachment.message')" rel="tooltip" data-placement="top">
-                  {{ $t('attachments.drawer.destination.attachment.message') }}</p>
-              </div>
-              <button :disabled="displayMessageDestinationFolder" class="buttonSelect d-flex pl-1 flex-column-reverse"
-                      @click="toggleSelectDestinationFolder()">
-                <i
-                  :title="!displayMessageDestinationFolder ? $t('attachments.drawer.destination.attachment') : $t('attachments.drawer.destination.attachment.access') "
-                  :class="displayMessageDestinationFolder ? 'disabled not-allowed' : ''" class="uiIconFolder " rel="tooltip"
-                  data-placement="top"></i>
-              </button>
-            </div>
-            <div class="uploadedFilesItems ml-5">
-              <transition-group name="list-complete" tag="div" class="d-flex flex-column">
-                <span
-                  v-for="attachment in attachments"
-                  :key="attachment"
-                  class="list-complete-item"
-                >
-                  <attachment-item :file="attachment"></attachment-item>
-                </span>
-              </transition-group>
-            </div>
-          </div>
         </div>
-        <attachments-folders-files-selector v-show="showDocumentSelector"
-                                            :is-cloud-enabled="isCloudDriveEnabled"
-                                            :mode-folder-selection="true"
-                                            :mode-folder-selection-for-file="modeFolderSelectionForFile"
-                                            :entity-id="entityId"
-                                            :entity-type="entityType"
-                                            :default-drive="defaultDrive"
-                                            :default-folder="defaultFolder"
-                                            @cancel="toggleServerFileSelector()"></attachments-folders-files-selector>
+        <attachments-drive-explorer-drawer
+          :is-cloud-enabled="isCloudDriveEnabled"
+          :mode-folder-selection="true"
+          :mode-folder-selection-for-file="modeFolderSelectionForFile"
+          :entity-id="entityId"
+          :entity-type="entityType"
+          :default-drive="defaultDrive"
+          :default-folder="defaultFolder"
+          @cancel="toggleServerFileSelector()"></attachments-drive-explorer-drawer>
         <div v-for="action in attachmentsComposerActions" :key="action.key" :class="`${action.appClass}Action`">
           <component v-dynamic-events="action.component.events" v-if="action.component"
                      v-bind="action.component.props ? action.component.props : {}"
                      :is="action.component.name" :ref="action.key"></component>
         </div>
       </template>
-      <template v-if="!showDocumentSelector" slot="footer">
+      <template slot="footer">
         <div class="d-flex align-center justify-space-between">
           <div class="limitMessage d-flex align-center grey--text">
             <i class="uiIconWarning my-auto pr-2 grey--text"></i>
@@ -177,7 +78,6 @@
 </template>
 
 <script>
-import axios from 'axios';
 import {getAttachmentsComposerExtensions} from '../js/extension';
 
 export default {
@@ -221,26 +121,18 @@ export default {
     return {
       showDestinationFolder: false,
       message: '',
-      uploadingFilesQueue: [],
       uploadingCount: 0,
-      maxUploadInProgressCount: 2,
       maxProgress: 100,
-      showDocumentSelector: false,
-      fileSizeLimitError: false,
-      filesCountLimitError: false,
-      sameFileError: false,
-      sameFileErrorMessage: `${this.$t('attachments.drawer.sameFile.error')}`,
-      BYTES_IN_MB: 1048576,
+      maxUploadInProgressCount: 2,
+      uploadingFilesQueue: [],
       MESSAGES_DISPLAY_TIME: 5000,
       drawerTitle: `${this.$t('attachments.upload.document')}`,
       pathDestinationFolder: '',
       showDestinationPath: false,
       schemaFolder: [],
       destinationFileName: '',
-      showDestinationFolderForFile: false,
       modeFolderSelectionForFile: false,
       showAttachmentsDrawer: false,
-      displayMessageDestinationFolder: true,
       cloudDriveConnecting: false,
       connectedDrive: {},
       privateFilesAttached: false,
@@ -272,21 +164,6 @@ export default {
     }
   },
   watch: {
-    fileSizeLimitError: function () {
-      if (this.fileSizeLimitError) {
-        setTimeout(() => this.fileSizeLimitError = false, this.MESSAGES_DISPLAY_TIME);
-      }
-    },
-    filesCountLimitError: function () {
-      if (this.filesCountLimitError) {
-        setTimeout(() => this.filesCountLimitError = false, this.MESSAGES_DISPLAY_TIME);
-      }
-    },
-    sameFileError: function () {
-      if (this.sameFileError) {
-        setTimeout(() => this.sameFileError = false, this.MESSAGES_DISPLAY_TIME);
-      }
-    },
     attachmentInfo: function () {
       if (this.attachmentInfo) {
         setTimeout(() => this.attachmentInfo = false, this.MESSAGES_DISPLAY_TIME);
@@ -301,8 +178,6 @@ export default {
         if (this.attachments.length === 0) {
           this.pathDestinationFolder = this.defaultDestinationFolderPath;
           this.schemaFolder = this.defaultSchemaFolder;
-          this.displayMessageDestinationFolder = true;
-          this.showDestinationPath = true;
         }
         this.privateFilesAttached = this.attachments.some(file => file.isPublic === false);
         this.fromAnotherSpaces = this.attachments.filter(({space}) => space && space.name !== this.groupId)
@@ -326,6 +201,7 @@ export default {
     }
   },
   created() {
+    document.addEventListener('paste', this.onPaste, false);
     this.$root.$on('remove-attachment-item', attachment => {
       this.removeAttachedFile(attachment);
     });
@@ -354,178 +230,10 @@ export default {
   methods: {
     openAttachmentsAppDrawer() {
       this.$refs.attachmentsAppDrawer.open();
-      this.$nextTick(() => {
-        if(!this.dragAndDropEventListenerInitialized){
-          ['drag', 'dragstart', 'dragend', 'dragover', 'dragenter', 'dragleave', 'drop'].forEach(function (evt) {
-
-            /*
-              For each event add an event listener that prevents the default action
-              (opening the file in the browser) and stop the propagation of the event (so
-              no other elements open the file in the browser)
-            */
-            document.addEventListener(evt, function (e) {
-              e.preventDefault();
-              e.stopPropagation();
-            }.bind(this), false);
-          }.bind(this));
-
-          document.addEventListener('dragover', function () {
-            this.$refs.dropFileBox.classList.add('dragStart');
-          }.bind(this));
-
-          /*
-            Capture the files from the drop event and add them to our local files
-            array.
-          */
-          this.$refs.dropFileBox.addEventListener('drop', function (e) {
-            this.$refs.dropFileBox.classList.remove('dragStart');
-            this.handleFileUpload(e.dataTransfer.files);
-          }.bind(this));
-
-          document.addEventListener('dragleave', function () {
-            this.$refs.dropFileBox.classList.remove('dragStart');
-          }.bind(this));
-
-          document.addEventListener('drop', function () {
-            this.$refs.dropFileBox.classList.remove('dragStart');
-          }.bind(this));
-
-          window.require(['SHARED/jquery'], function ($) {
-            $('#exoAttachmentsApp *[rel="tooltip"]').tooltip();
-          });
-          this.dragAndDropEventListenerInitialized = true;
-        }
-      });
-
     },
     closeAttachmentsAppDrawer() {
+      document.removeEventListener('paste', this.onPaste, false);
       this.$refs.attachmentsAppDrawer.close();
-    },
-    uploadFile: function () {
-      this.$refs.uploadInput.click();
-    },
-    handleFileUpload: function (files) {
-      const newFilesArray = Array.from(files);
-
-      newFilesArray.sort(function (file1, file2) {
-        return file1.size - file2.size;
-      });
-
-      const newAttachedFiles = [];
-      newFilesArray.forEach(file => {
-        newAttachedFiles.push({
-          originalFileObject: file,
-          fileDrive: this.currentDrive,
-          name: file.name,
-          size: file.size,
-          mimetype: file.type,
-          uploadId: this.getNewUploadId(),
-          uploadProgress: 0,
-          destinationFolder: this.pathDestinationFolder,
-          pathDestinationFolderForFile: '',
-          isPublic: true
-        });
-      });
-
-      this.uploadMode = 'temp';
-      newAttachedFiles.forEach(newFile => {
-        this.queueUpload(newFile);
-      });
-      this.$refs.uploadInput.attachments = '';
-    },
-    getFormattedFileSize(fileSize) {
-      const formattedSizePrecision = 2;
-      const sizeMB = fileSize / this.BYTES_IN_MB;
-      return sizeMB.toFixed(formattedSizePrecision);
-    },
-    getNewUploadId: function () {
-      const maxUploadId = 100000;
-      return Math.floor(Math.random() * maxUploadId);
-    },
-    queueUpload: function (file) {
-      if (this.uploadMode === 'temp') {
-        if (this.attachments.length >= this.maxFilesCount) {
-          this.filesCountLimitError = true;
-          return;
-        }
-
-        const fileSizeInMb = file.size / this.BYTES_IN_MB;
-        if (fileSizeInMb > this.maxFileSize) {
-          this.fileSizeLimitError = true;
-          return;
-        }
-        const fileExists = this.attachments.some(f => f.name === file.name);
-        if (fileExists) {
-          this.sameFileErrorMessage = this.sameFileErrorMessage.replace('{0}', file.name);
-          this.sameFileError = true;
-          return;
-        }
-
-        this.attachments.push(file);
-      }
-      if (this.uploadingCount < this.maxUploadInProgressCount) {
-        if (this.uploadMode === 'temp') {
-          this.sendFileToServer(file);
-        } else {
-          this.uploadFileToDestinationPath(file);
-        }
-      } else {
-        this.uploadingFilesQueue.push(file);
-      }
-    },
-    processNextQueuedUpload: function () {
-      if (this.uploadingFilesQueue.length > 0) {
-        if (this.uploadMode === 'temp') {
-          this.sendFileToServer(this.uploadingFilesQueue.shift());
-        } else {
-          this.uploadFileToDestinationPath(this.uploadingFilesQueue.shift());
-        }
-      }
-    },
-    sendFileToServer: function (file) {
-      this.uploadingCount++;
-      this.$emit('uploadingCountChanged', this.uploadingCount);
-
-      const formData = new FormData();
-      formData.append('file', file.originalFileObject);
-
-      const uploadUrl = `${eXo.env.server.context}/upload?action=upload&uploadId=${file.uploadId}`;
-      // Had to use axios here since progress observation is still not supported by fetch
-      axios.request({
-        method: 'POST',
-        url: uploadUrl,
-        credentials: 'include',
-        data: formData,
-        onUploadProgress: (progress) => {
-          file.uploadProgress = Math.round(progress.loaded * this.maxProgress / progress.total);
-        }
-      }).then(() => {
-
-        // Check if the file has correctly been uploaded (progress=100) before refreshing the upload list
-        const progressUrl = `${eXo.env.server.context}/upload?action=progress&uploadId=${file.uploadId}`;
-        fetch(progressUrl)
-          .then(response => response.text())
-          .then(responseText => {
-            // TODO fix malformed json from upload service
-            let responseObject;
-            try {
-              // trick to parse malformed json
-              eval(`responseObject = ${responseText}`); // eslint-disable-line no-eval
-            } catch (err) {
-              return;
-            }
-
-            if (!responseObject.upload[file.uploadId] || !responseObject.upload[file.uploadId].percent ||
-              responseObject.upload[file.uploadId].percent !== this.maxProgress.toString()) {
-              this.removeAttachedFile(file.uploadId);
-            } else {
-              file.uploadProgress = this.maxProgress;
-            }
-          });
-        this.uploadingCount--;
-        this.$emit('uploadingCountChanged', this.uploadingCount);
-        this.processNextQueuedUpload();
-      });
     },
     uploadFileToDestinationPath: function (file) {
       this.uploadingCount++;
@@ -579,26 +287,12 @@ export default {
     addDestinationFolderForAll(pathDestinationFolder, folder, currentDrive) {
       this.currentDrive = currentDrive;
       this.pathDestinationFolder = pathDestinationFolder;
-      if (pathDestinationFolder === '') {
-        this.showDestinationPath = false;
-      } else {
-        this.showDestinationPath = true;
-      }
+      this.schemaFolder = folder.split('/');
       for (let i = 0; i < this.attachments.length; i++) {
         if (!this.attachments[i].destinationFolder || this.attachments[i].destinationFolder === this.defaultDestinationFolderPath) {
           this.attachments[i].destinationFolder = this.pathDestinationFolder;
           this.attachments[i].fileDrive = this.currentDrive;
         }
-      }
-      this.schemaFolder = [];
-      const namesOfFolders = folder.split('/');
-      for (let i = 0; i < namesOfFolders.length; i++) {
-        this.schemaFolder[i] = namesOfFolders[i];
-      }
-      this.showDocumentSelector = !this.showDocumentSelector;
-      this.drawerTitle = this.showDocumentSelector ? `${this.$t('attachments.drawer.existingUploads')}` : `${this.$t('attachments.upload.document')}`;
-      if (!this.showDocumentSelector) {
-        this.showDestinationFolder = false;
       }
     },
     addDestinationFolderForFile(pathDestinationFolder, folder, isPublic, currentDrive) {
@@ -611,11 +305,6 @@ export default {
           this.attachments[i].isPublic = isPublic;
         }
       }
-      this.showDocumentSelector = !this.showDocumentSelector;
-      this.drawerTitle = this.showDocumentSelector ? `${this.$t('attachments.drawer.existingUploads')}` : `${this.$t('attachments.upload.document')}`;
-      if (!this.showDocumentSelector) {
-        this.showDestinationFolderForFile = false;
-      }
       this.modeFolderSelectionForFile = false;
     },
     toggleServerFileSelector(selectedFiles) {
@@ -625,46 +314,15 @@ export default {
         this.$emit('input', this.attachments);
         this.$emit('attachmentsChanged', this.attachments);
       }
-      this.showDocumentSelector = !this.showDocumentSelector;
-      this.drawerTitle = this.showDocumentSelector ? `${this.$t('attachments.drawer.existingUploads')}` : `${this.$t('attachments.upload.document')}`;
-      if (!this.showDocumentSelector) {
-        this.showDestinationFolder = false;
-        this.showDestinationFolderForFile = false;
-      }
-    },
-    toggleSelectDestinationFolder() {
-      this.showDestinationFolder = true;
-      this.showDocumentSelector = !this.showDocumentSelector;
-      this.drawerTitle = this.showDocumentSelector ? `${this.$t('attachments.drawer.destination.folder')}` : `${this.$t('attachments.upload.document')}`;
     },
     openSelectDestinationFolderForFile(file) {
       this.modeFolderSelectionForFile = true;
       this.destinationFileName = file.name;
-      this.showDestinationFolderForFile = true;
-      this.showDocumentSelector = !this.showDocumentSelector;
-      this.drawerTitle = this.showDocumentSelector ? `${this.$t('attachments.drawer.destination.folder')}` : `${this.$t('attachments.upload.document')}`;
-    },
-    addDefaultPath() {
-      if (eXo.env.portal.spaceId) {
-        this.$attachmentsService.getSpaceById(eXo.env.portal.spaceId).then(space => {
-          this.schemaFolder.push(space.displayName);
-          this.schemaFolder.push('Activity Stream Documents');
-          this.showDestinationPath = true;
-          this.isActivityStream = false;
-          this.spaceGroupId = space.groupId;
-        });
-      } else {
-        this.schemaFolder.push(eXo.env.portal.userName);
-        this.schemaFolder.push('Personal Documents');
-        this.schemaFolder.push('Documents');
-        this.showDestinationPath = true;
-        this.isActivityStream = true;
-      }
+      this.$root.$emit('open-drive-explorer-drawer');
     },
     deleteDestinationPathForFile(fileName) {
       for (let i = 0; i < this.attachments.length; i++) {
         if (this.attachments[i].name === fileName) {
-          this.attachments[i].showDestinationFolderForFile = '';
           this.attachments[i].pathDestinationFolderForFile = '';
           this.attachments[i].fileDrive = this.currentDrive;
           this.attachments[i].destinationFolder = this.pathDestinationFolder;
@@ -687,7 +345,7 @@ export default {
           this.getFilename(textItem, textItemType, name => {
             //needed to create a new file to be able to rename it.
             const myNewFile = new File([pastedFile], name, {type: pastedFile.type});
-            this.handleFileUpload([myNewFile]);
+            this.$root.$emit('handle-pasted-files-from-clipboard',[myNewFile]);
           });
         }
       }
@@ -737,8 +395,19 @@ export default {
         this.closeAndResetAttachmentsDrawer();
       }
     },
+    queueUpload(file) {
+      if (this.uploadingCount < this.maxUploadInProgressCount) {
+        this.uploadFileToDestinationPath(file);
+      } else {
+        this.uploadingFilesQueue.push(file);
+      }
+    },
+    processNextQueuedUpload: function () {
+      if (this.uploadingFilesQueue.length > 0) {
+        this.uploadFileToDestinationPath(this.uploadingFilesQueue.shift());
+      }
+    },
     initDefaultDestinationFolderPath(defaultDestinationFolderPath, folderName) {
-      this.showDestinationPath = true;
       this.defaultDestinationFolderPath = defaultDestinationFolderPath || '';
       if (folderName) {
         this.defaultSchemaFolder = [];
