+++ ecm-wcm-extension/src/main/webapp/attachments-app/components/attachments-drive-explorer/AttachmentsFolderActionsMenu.vue
